move meson create exe generator
